requirements.txt
streamlit==1.49.1
streamlit-webrtc==0.47.7
av==10.0.0
pydub==0.25.1
numpy==1.26.4
faster-whisper==1.0.3
google-generativeai==0.7.2
python-docx==1.1.2
audio-recorder-streamlit==0.0.8